Adds logout functionality as a reusable hook
Implements a `useLogout` hook to encapsulate logout logic, including clearing user data from Redux, displaying a success message, and navigating to the login page. This promotes code reusability and maintainability by centralizing the logout process. The hook is then used in the Sidebar and Navbar components, removing the need for duplicate code.

Also, corrects the path used to find a user by email in Firebase.